+++ ChangeLog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Upgraded nlohmann::json to version 3.11.2
+- Removed json from the global namespace and moved nlohmann::json to the ifm3d::json namespace
 
 ### Fixes 
 - Missing python bindings for `O3R::Port` and `O3R::Ports`